Update PropTypes and Fix Bug iOS Code
PropTypes ->
In the new versions of React Native, PropTypes is no longer working when imported directly from React, as errors occur. For this reason the prop-types library was added and the import was changed.

Bug Fix iOS ->
When I compiled the code for iOS an error was reported, so I also changed this error based on another commit:

https://github.com/yz1311/react-native-smart-barcode/commit/82ceb0aa12adf96a121349ef9e606390a7a79153

https://github.com/yz1311/react-native-smart-barcode/commit/acb89ecab8a4b81006bcc05c4290e51947e7cd60